kafka rebalancing sdc

+++ scripts/docker_cleanup_all.sh
@@ -1,9 +1,25 @@
 #!/bin/bash
 
-docker stop $(docker ps -aq)
-docker rm $(docker ps -aq)
+numofps=`docker ps -aq | wc  -l`
+if [ $numofps -ne 0 ]
+then
+    docker stop $(docker ps -aq)
+fi
+echo "docker stopped all containers."
 
-echo "docker stopped and removed all containers."
+numofps=`docker ps -aq | wc  -l`
+if [ $numofps -ne 0 ]
+then
+    docker rm $(docker ps -aq)
+fi
+echo "docker removed all containers."
 
+numofps=`docker volume ls -q | wc  -l`
+if [ $numofps -ne 0 ]
+then
+    docker volume rm $(docker volume ls -q)
+fi
+echo "docker removed all volumes."
 
-#docker network prune -f
+docker network prune -f
+echo "docker pruned all networks."
